electron: Move kolibri environment to electron app
This patch moves all the environment variables calculation to the
kolibri-electron, so now the content fallback, the kolibri-home, the
pythonpath for extension and the copy of the preseeded home is done in
the kolibri-electron app.

With this change we've more control over the initialization so we can
show more information in the loading screen.

https://phabricator.endlessm.com/T32261