CI: Improve build scripts
 * Use latest kolibri stable (0.15.0)
 * Remove kolibri modifications, not needed for 0.15.0
 * Use yarn cache
 * Create a ZIP file and upload to tag if it's a release

The electron app fails to build on windows with large paths and that's
hapenning if we copy the Kolibri folder inside the resources of the
electron app before the build.

This patch does the copy of the Kolibri folder built with kapew after
the electron is built so it won't fail.

The zip compression before uploading as artifact is also a good practice
because when there are a lot of files it could fail, and we can also
upload the zip file automatically to the release if it's a tag.

https://phabricator.endlessm.com/T32867