Bringing in Windows package manifest and pipeline
Bring the appxmanifest.xml for APPX/MSIX packaging. Also, add the steps
into the CI pipeline to build the package and upload the artifact.

The first 3 fields of Identity's Version in appxmanifest.xml should be
same as the release version. The fourth field could be build times, or
anything else.

https://phabricator.endlessm.com/T33442